Integrate change 271128 from ZDESKTOP-700        bug: 55474
       Add IOException message skipping for POP and IMAP. Individual 
       messages are skipped and tried again later if they cannot be saved 
       due to IOException.

http://bugzilla.zimbra.com/show_bug.cgi?id=55474

+++ ZimbraOffline/src/java/com/zimbra/cs/mailbox/SyncFailure.java
@@ -15,16 +15,16 @@
 package com.zimbra.cs.mailbox;
 
 public class SyncFailure {
-    private int itemId;
+    private long itemId;
     private Exception exception;
     private String message;
-    public SyncFailure(int itemId, Exception exception, String message) {
+    public SyncFailure(long itemId, Exception exception, String message) {
         super();
         this.itemId = itemId;
         this.exception = exception;
         this.message = message;
     }
-    public int getItemId() {
+    public long getItemId() {
         return itemId;
     }
     public Exception getException() {
